requirements.txt
# Core MCP dependencies
mcp
mcp-use
psutil
pathlib2
typing-extensions
python-dotenv
requests
ollama
langchain-ollama
websockets
openai
panda
lancedb
sentence-transformers
tzdata